-- Main module (prep_module) --
Do the practice with :
cloning objects,
creating objects using serialization/deserialization,
functional interfaces and lambda expressions,
with the cast() operation.

-- Additional module (test-new-playwright) --

Do the practice with :
create two tests for valid and invalid login to the system,
create the necessary classes for the pages and components,
create a common configuration based on the Owner dependency,
create a factory for switching between pages.